[redhat] config review: cleanup 9P_FS options
As noted by Steven Whitehouse on config review, 9P isn't needed. It was already
disabled. This commit just cleanups the setting of 9P_FS options: as
CONFIG_NET_9P is already disabled and 9P_FS depends on it, we don't need to
disable them explicitly as they are already unselectable.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>