[redhat] config review: mark CONFIG_FS_ENCRYPTION=n
As noted on config review by Steven Whitehouse, we should leave it disabled. It
was already being done by default, just mark and add the corresponding config
file, no change in practice. And anyway, this option is usually auto-selected,
but no one requires it at the moment with current config.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>