[redhat] config review: cleanup fs related options
There are a lot of options that we were trying to set, but they are gone already
or got hidden under another dependency/option that we already disable. Also,
some options are auto selected and doesn't matter what we try to choose. For all
these cases, remove them from the tree. I verified manually the generated
configs before/after and there are no changes after this cleanup (no diff).

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>