[redhat] config review: set CONFIG_MLXSW_CORE=m on aarch64 and x86_64
This change is based on feedback from config review by Mark Langsdorf,
Jarod Wilson, Don Dutile, John Linville. It was also commented that the
driver being enabled should be marked as 'Tech Preview', however this is
out of scope of config changes and must be done by someone else later.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>