[redhat] config review: cleanup CONFIG_CRASH settings
There are two issues at the moment: all arches enable CONFIG_CRASH
(where it could be done only with one file in the generic directory),
and some enable it as a module, others builtin. Since crash utility is
able to load the module when necessary, just set CONFIG_CRASH=m once in
the generic directory.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>