[redhat] config review: cleanup some GPIO options
Some options were removed, or just depends on a setting which is already
disabled. Also, consolidate one or other setting which was being set to
the same value in all arches, with just one file in the generic
directory.

There is one special case here, which is CONFIG_GPIO_LANGWELL that was
renamed to CONFIG_GPIO_INTEL_MID, but as we don't enable
CONFIG_X86_INTEL_MID, we can't set GPIO_INTEL_MID anyway, thus the
config files were just removed in this case.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>