[redhat] config review: set CONFIG_SENSORS_I5500=m on x86_64
This seems an obvious choice, since RHEL etc. supports Intel
5500/5520/X58 chipset/platform, thus enable the Intel 5500/5520/X58
temperature sensor (even though pegas on x86_64 is for testing only at the
moment).

There was no feedback specific to this option in config review, however
the feedback/"feelings" from Dean Nelson/Jarod Wilson were to enable sensors
unless it is "some embedded platform or SoC we don't want to support". This to
me falls into the category of supported platform where we want the sensor
to be enabled.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>