[redhat] add support for aarch64 in rh-cross-* make targets
Message-id: <20170117230937.15339-1-ahs3@redhat.com>
Patchwork-id: 162692
O-Subject: [Pegas 1.0 PATCH v2] redhat: add support for aarch64 in rh-cross-* make targets
Bugzilla: 1415855
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>
RH-Acked-by: Jonathan Toppins <jtoppins@redhat.com>

This is an addition to the Red Hat makefiles that I find convenient; it
may or may not work with stock RHEL, and is not intended for use in building
a released kernel since those must be done with native compilers -- it is
only a convenience for trying out different kernel changes.

Pegas already supported cross-builds for a couple of other architectures
when running on an x86 system.  This provides the proper targets so that

   $ make rh-cross-aarch64-rpms

will also work to build a kernel using aarch64 cross compilers.  Since
this involved a little re-factoring of Makefile.cross, the architecture
specific targets are also in alphabetical order now.

It is also possible now to only cross build a single architecture at a
time (the original default of all cross builds remains the default).  For
example:

   $ make rh-cross-download

checks for all cross-compilers, while:

   $ ARCH=arm64 make rh-cross-download

will only check for aarch64 cross-compilers.  This also allows:

   $ ARCH=arm64 make rh-cross-aarch64-rpms

so that one does not have to install cross tools for all architectures
to only build one.

Signed-off-by: Al Stone <ahs3@redhat.com>

v2:
   -- proper formatting in email
   -- fix to conditional in x86_rngd.sh suggested by Jonathan Toppins

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>