[redhat] Fix some cross-compilation issues
Message-id: <20170621224800.25547-2-ahs3@redhat.com>
Patchwork-id: 180057
O-Subject: [[Pegas1.0 BZ1460827 v5 PATCH 1/3] Fix some cross-compilation issues
Bugzilla: 1460827
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

In the process of patching for bz1415855, I missed a couple of items
that ended up making cross-compiling more restrictive than it needs to
be -- in particular, one is currently required to only use RHEL cross
tools and specific paths.  The changes with this patch allow one to
use the ARCH and CROSS_COMPILE variables to use whatever compiler path
one chooses.  Further, the logic for determining whether or not a
cross-environment is reasonable has been moved to the single script
redhat/scripts/download_cross.sh instead of being partially there
and in the Makefile.

Signed-off-by: Al Stone <ahs3@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>