[redhat] Makefile.cross: cleaned up inconsistent use of s390/s390x
Message-id: <20170621224800.25547-4-ahs3@redhat.com>
Patchwork-id: 180058
O-Subject: [[Pegas1.0 BZ1460827 v5 PATCH 3/3] [redhat] redhat/Makefile.cross: cleaned up inconsistent use of s390/s390x
Bugzilla: 1460827
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

In Makefile.cross, a couple of the instances where generate_cross_report.sh
was invoked for s390x used "s390" instead of "s390x".  With the addition
of a report on the cross-compiler being used, this would cause the cross
report to fail to find the s390x cross-compilers, even if installed.

This patch changes the "s390" to "s390x" so that it is now consistent
throughout Makefile.cross.

Signed-off-by: Al Stone <ahs3@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>