[redhat] Ignore localversion and configs directory
Message-id: <1498507385-2270-1-git-send-email-dzickus@redhat.com>
Patchwork-id: 180163
O-Subject: [RHEL-ALT PEGAS PATCH] [redhat] Ignore localversion and configs directory
Bugzilla:
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Jonathan Toppins <jtoppins@redhat.com>

Recently a developer new to the Red Hat build process accidentally
the localversion file, brew built it and watched it failed to boot.

I am not entirely sure why this fails but it has to do with us
setting the localversion twice, once in the spec file, then accidentally
the second time with the saved 'localversion' file.

RHEL-5/6/7 prevented this by putting 'localversion' in the .gitignore
file.  So if you try to 'git add' it, it flags you and makes you
think twice before doing so.

I also noticed, rhel-5/6/7 threw the configs/ in .gitignore too, add
those as well.  This should sync up some of the .gitignore file.

Signed-off-by: Don Zickus <dzickus@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>