[powerpc] Change message for IBM Power 8 unsupported
Message-id: <20170712165620.25994.30097.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 180519
O-Subject: [PATCH PEGAS1.0 BZ1470052] Change message for IBM Power 8 unsupported
Bugzilla: 1470052
RH-Acked-by: Peter Martuccelli <peterm@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1470052

Description:
------------
I added a message that IBM Power 8 isn't supported. The message included the variant
of Power 8 and the variant really doesn't add anything to the message, so remove
the variant from the message.

Brew:
-----
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=13640277

Upstream:
---------
This is a RHEL only change.

Test Status:
------------

on Power 9 system the msg isn't displayed and that is what we want

before the patch on power 8 system (ibm-p8-rhevm-09.lab4)

[root@ibm-p8-rhevm-09 /]# dmesg | grep  "Detected Power 8"
[    0.000000] Detected Power 8 E processor

after the patch

[root@ibm-p8-rhevm-09 ~]# dmesg | grep  "Detected Power 8"
[    0.000000] Detected Power 8 processor

----------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>