[ata] ahci: thunderx2: Fix for errata that affects stop engine
Message-id: <1502805127-8211-3-git-send-email-rrichter@redhat.com>
Patchwork-id: 182720
O-Subject: [RHEL-Alt-7.4 BZ 1430391 v6 2/2] ahci: thunderx2: Fix for errata that affects stop engine
Bugzilla: 1430391
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Jon Masters <jcm@redhat.com>
RH-Acked-by: Jonathan Toppins <jtoppins@redhat.com>

From: Jayachandran C <jnair@caviumnetworks.com>

RHEL only, early silicon enablement, there is a hw fix for B0 production
silicon.

v6:
 * fix failures during continuous reboot test

v5:
 * updated description, added acks, no code changes

v4:
 * added check to restict to pci devs on the soc only

Apply workaround for this errata:
  Synopsis: Resetting PxCMD.ST may hang the SATA device

  Description: An internal ping-pong buffer state is not reset
  correctly for an PxCMD.ST=0 command for a SATA channel. This
  may cause the SATA interface to hang when a PxCMD.ST=0 command
  is received.

  Workaround: A SATA_BIU_CORE_ENABLE.sw_init_bsi must be asserted
  by the driver whenever the PxCMD.ST needs to be de-asserted. This
  will reset both the ports. So, it may not always work in a 2
  channel SATA system.

  Resolution: Fix in B0.

Add the code to ahci_stop_engine() to do this. It is not easy to
stop the other "port" since it is associated with a different AHCI
interface. Please note that with this fix, SATA reset does not
hang any more, but it can cause failures on the other interface
if that is in active use.

Unfortunately, we have nothing other the the CPU ID to check if the
SATA block has this issue.

Signed-off-by: Jayachandran C <jnair@caviumnetworks.com>
[added check to restict to pci devs on the soc only]
[fix failures during continuous reboot test]
Signed-off-by: Robert Richter <rrichter@cavium.com>
Signed-off-by: Robert Richter <rrichter@redhat.com>

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>