[redhat] config: re-enable bnx2 driver and mark unsupported
Message-id: <4fa66f466546b38856ea619be3d4cabcf4c997af.1505061391.git.jtoppins@redhat.com>
Patchwork-id: 184595
O-Subject: [RHEL-ALT-7.4 2/3] [redhat] config: re-enable bnx2 driver and mark unsupported
Bugzilla: 1481788
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>

Enabling this driver at the request of Quality Engineering(QE) to support
testing on x86 based systems of which we have over 600 systems in beaker.
This driver is not intended to be used by customers and is not listed as
a supported card for any server configuration supported by the Pegas
program. Hence, enable the driver to allow for internal testing and
mark this driver as unsupported and taint the kernel when such a
device supported by this driver is attempted to be used.

Since newer kernels use a generic function to register pci based drivers
there is no easy way to taint the kernel upon module loading. Instead
the kernel will be tained if a device is attempted to be probed by the
bnx2 driver.

Upstream Status: RHEL Only
Bugzilla: 1481788
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=13971480
Tested: built and verified bnx2.ko existed in the kernel rpm, also verified
        the kernel was tained upon attempting to have the driver configure
        a bnx2 based device. The kernel is not tainted if the module is
        simply loaded and no bnx2 cards exist in the system.

Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>