[powerpc] powerpc/powernv: Rework EEH initialization on powernv
Message-id: <20170928172209.13510.80581.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 186997
O-Subject: [PATCH RHEL-ALT-7.4 BZ1496734 4/4] powerpc/powernv: Rework EEH initialization on powernv
Bugzilla: 1496734
RH-Acked-by: David Gibson <dgibson@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1496734

Description:
------------
Remove the post_init callback which is only used
by powernv, we can just call it explicitly from
the powernv code.

This partially kills the ability to "disable" eeh at
runtime via debugfs as this was calling that same
callback again, but this is both unused and broken
in several ways. If we want to revive it, we need
to create a dedicated enable/disable callback on the
backend that does the right thing.

Let the bulk of eeh initialize normally at
core_initcall() like it does on pseries by removing
the hack in eeh_init() that delays it.

Instead we make sure our eeh->probe cleanly bails
out of the PEs haven't been created yet and we force
a re-probe where we used to call eeh_init() again.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Acked-by: Russell Currey <ruscur@russell.cc>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>

Upstream Status:
----------------
commit b9fde58db7e5738cacb740b0ec547933fe314fbe

https://git.kernel.org/pub/scm/linux/kernel/git/powerpc/linux.git/commit/arch/powerpc?h=next&id=b9fde58db7e5738cacb740b0ec547933fe314fbe

----------------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>