[arm64] DO NOT UPSTREAM - pmuv3: disable PMUv3 in VM when vPMU=off
Message-id: <20171013161738.5644-2-drjones@redhat.com>
Patchwork-id: 193242
O-Subject: [RHEL-ALT-7.5 BZ1501452 PATCH 1/1] [arm64] DO NOT UPSTREAM - pmuv3: disable PMUv3 in VM when vPMU=off
Bugzilla: 1501452
RH-Acked-by: Al Stone <ahs3@redhat.com>
RH-Acked-by: Wei Huang <wei@redhat.com>

Current PMUv3 code checks the existence of PMUv3 by reading the PMU_VER
field of ID_AA64DFR0_EL1. This works pretty well on bare-metal machines.
However, we found that ID_AA64DFR0_EL1 isn't intercepted by KVM. It
means that PMUv3, running inside a VM, always reads the host value of
ID_AA64DFR0_EL1 even when vPMU=off. So PMUv3 will always be loaded
inside a VM if the host machine supports PMUv3. This apparently is not
what we expect. Instead we want PMUv3 be disabled completely when
vPMU=off.

This patch addresses the problem by checking the IMP_ID field of
PMCR. This approach works because PMCR's IMP_ID field returns 0 when
vPMU=off. With this patch, QEMU's pmu=on/off can correctly control PMUv3
inside a VM. It's safe for bare-metal, as no HW vendors should being
using 0 for their IMP_ID.

NOTE: this patch isn't a complete solution, that is why we didn't
submit it to the upstream. A complete fix requires ID_AA64DFR0_EL1 been
intercepted by KVM. But that is an extremely big patch because many other
ID registers need to be intercepted and handled as well. According to
Jeremy Linton, such patch is being prepared by ARM internally. But we
can't wait for it for 7.5.

This is a forward port of kernel-alt 7.4 commit 67ff8b0ce185.

Cc: Wei Huang <wei@redhat.com>
Signed-off-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>