[arm64] DO NOT UPSTREAM - topology: Adjust sysfs topology
Message-id: <20171013161839.5861-2-drjones@redhat.com>
Patchwork-id: 193244
O-Subject: [RHEL-ALT-7.5 BZ1501443 PATCH 1/1] [arm64] DO NOT UPSTREAM - topology: Adjust sysfs topology
Bugzilla: 1501443
RH-Acked-by: Mark Langsdorf <mlangsdo@redhat.com>
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Al Stone <ahs3@redhat.com>
RH-Acked-by: Jonathan Toppins <jtoppins@redhat.com>

Extracting topology information from the MPIDR Affx fields currently
generates improper information under sysfs with upstream Linux. DT has
a format, if provided, albeit that is under an RFC as part of the arm64
NUMA support, and wouldn't work for RHELSA.  ACPI is developing new
structures to provide the proper info, but is a bit off before its
arrival. This patch is a placeholder that should work for the known
ARM64 SoCs expected to be supported in RHELSA over the next release
cycle.

Note: On x86, the vast majority of this information is provided
via CPUID invocations, and for really large systems, partial ACPI
information.  There is no such equivalent on arm64 yet.

The patch adds a new function that is run at late_initcall() time
which scans the MPIDR.Affx fields that were read and stored for each
CPU as they were brought online. The function must wait for all CPUs
to be online as the MPIDR is per-CPU and doesn't know the number of
clusters (Aff1, Aff2 values) until after all CPUs have been seen.
late_initcall() is after all CPUs are online and before sysfs can be
mounted.

This patch also supports parsing the MPIDR of multi-socket systems.  The
socket ID is in the next higher affinity level than the cluster ID, and
we use it to populate physical_package_id. We still effectively remove
cluster_id as it's too late to expose that information for 7.5. We can't
leave the zeroing of it, though, because update_siblings_masks() uses it
alone to determine core siblings. I.e. cluster_id needs to be an ID that
is unique across sockets, otherwise we'll get inter-socket siblings. We
can just use the extracted socket ID for cluster_id to solve that.

multi-socketed and multi-threaded CPUs also show up correctly for the
tested platforms with lstopo by using Aff0 for the thread-id of a multi-
threaded system and then shifting the affinity level used up for the
other identifiers.

This is a squashed forward port from RHELSA 7.4. The 7.4 commits are:

  d441f38a1949 [arm64] topology: Adjust sysfs topology
  159b70387fbf [arm64] aarch64: topology: Fix CPU topology description on multi-socket platforms

Cc: Jonathan Toppins <jtoppins@redhat.com>
Cc: Al Stone <ahs3@redhat.com>
Signed-off-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>