[powerpc] Add PPC_FEATURE2_HTM_NO_SUSPEND
Message-id: <20171027215813.31340.63019.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 195342
O-Subject: [PATCH RHEL-ALT-7.5 BZ1498555 2/5] powerpc: Add PPC_FEATURE2_HTM_NO_SUSPEND
Bugzilla: 1498555
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1498555

Description:
------------
Some CPUs can operate in a mode where TM (Transactional Memory) is
enabled but the suspended state of TM is disabled. In this mode
tsuspend does not enter suspended state, instead the transaction is
aborted. Similarly any other event that would lead to suspended state
instead aborts the transaction.

There is also an ABI change, in that in this mode processes are not
allowed to sigreturn with an MSR that would lead to suspended state,
Linux will instead return an error to the sigreturn syscall.

Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>

Upstream Status:
----------------
commit cba6ac4869e45cc93ac5497024d1d49576e82666

https://git.kernel.org/pub/scm/linux/kernel/git/powerpc/linux.git/commit/?id=cba6ac4869e45cc93ac5497024d1d49576e82666

----------------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>