[powerpc] powerpc/powernv: Enable TM without suspend if possible
Message-id: <20171027215819.31340.12069.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 195345
O-Subject: [PATCH RHEL-ALT-7.5 BZ1498555 3/5] powerpc/powernv: Enable TM without suspend if possible
Bugzilla: 1498555
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1498555

Description:
------------
Some Power9 revisions can run in a mode where TM operates without
suspended state. If we find ourself on a CPU that might be in this
mode, we query OPAL to check, and if so we reenable TM in CPU
features, and enable a new user feature to signal to userspace that we
are in this mode.

We do not enable the "normal" user feature, PPC_FEATURE2_HTM, but we
do enable PPC_FEATURE2_HTM_NOSC because that indicates to userspace
that the kernel will abort transactions on syscall entry, which is
true regardless of the suspend mode.

Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>

Upstream Status:
----------------
commit 54820530c5faa9fd78e1c08cb6449100b1a19157

https://git.kernel.org/pub/scm/linux/kernel/git/powerpc/linux.git/commit/?id=54820530c5faa9fd78e1c08cb6449100b1a19157

----------------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>