[netdrv] ibmvnic: Update reset infrastructure to support tunable parameters
Message-id: <20171101204707.42970.10783.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 195467
O-Subject: [PATCH RHEL-ALT-7.5 BZ1464528 4/5] ibmvnic: Update reset infrastructure to support tunable parameters
Bugzilla: 1464528
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1464528

Description:
------------
Update ibmvnic reset infrastructure to include a new reset option that will
allow changing of tunable parameters. There currently is no way to request
different capabilities from the vnic server on the fly so this patch
achieves this by resetting the driver and attempting to log in with the
requested changes. If the reset operation fails, the old values of the
tunable parameters are stored in the "fallback" struct and we attempt to
login with the fallback values.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Upstream Status:
----------------
commit c26eba03e4073bd32ef6c0ea2ba2a3ff5eed11da

https://git.kernel.org/pub/scm/linux/kernel/git/davem/net-next.git/commit/?id=c26eba03e4073bd32ef6c0ea2ba2a3ff5eed11da

----------------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>