[netdrv] ibmvnic: Fix failover error path for non-fatal resets
Message-id: <20171101204719.42970.2840.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 195468
O-Subject: [PATCH RHEL-ALT-7.5 BZ1464528 5/5] ibmvnic: Fix failover error path for non-fatal resets
Bugzilla: 1464528
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1464528

Description:
------------
For all non-fatal reset conditions, the hypervisor will send a failover when
we attempt to initialize the crq and the vnic client is expected to handle
that failover instead of the existing non-fatal reset. To handle this, we
need to return from init with a return code that indicates that we have hit
this case.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Upstream Status:
----------------
commit 2a1bf51111975846f412f47449edefdf6fa17ee4

https://git.kernel.org/pub/scm/linux/kernel/git/davem/net-next.git/commit/?id=2a1bf51111975846f412f47449edefdf6fa17ee4

----------------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>