[block] sg_io: introduce unpriv_sgio queue flag
Message-id: <20171109182547.32400-3-pbonzini@redhat.com>
Patchwork-id: 197464
O-Subject: [RHEL7.5 kernel-alt PATCH 2/3] sg_io: introduce unpriv_sgio queue flag
Bugzilla: 1492769
RH-Acked-by: Ming Lei <ming.lei@redhat.com>
RH-Acked-by: Ewan Milne <emilne@redhat.com>

This queue flag will let unprivileged users send any SG_IO command to the
device, without any filtering.  This makes it possible to run a program
where you want to access the full range of SCSI commands, while still
running as confined as possible.  With this patch, such a program will
not need the CAP_SYS_RAWIO capability anymore.  It will also *not* be
able to send SCSI commands to a partition, or to a device mapper target
that does not map the entirety of the underlying block device.

The RHEL7 BZ had another patch, which has been resubmitted independently
and is now upstream (commit 25cdb64, "block: allow WRITE_SAME commands
with the SG_IO ioctl", 2016-12-19) after having been rejected when I
submitted it in 2012.  Because of this, I am going to try again getting
the patch upstream for 4.12.  However, it will hopefully have a more
powerful (but different) userspace API.  So for backwards compatibility
reasons Pegas is stuck with this non-upstream patch.

Cc: Jeff Moyer <jmoyer@redhat.com>
Cc: Ewan Milne <emilne@redhat.com>
Cc: Ming Lei <ming.lei@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>

Additional note from Paolo Bonzini related to this patch/patchset:

These three patches are not upstream.  For RHEL8 we can choose between
trying again to push the patches upstream, or use QEMU's new ability to
send persistent reservations commands through a privileged daemon.  In
any case for RHEL7 we're stuck with it because it's part of the expected
userspace API.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>