[cpufreq] stats: Handle the case when trans_table goes beyond PAGE_SIZE
Message-id: <1511020300-2033-1-git-send-email-drosario@redhat.com>
Patchwork-id: 199237
O-Subject: [RHEL-ALT-7.5 BZ1499411 PATCH] cpufreq: stats: Handle the case when trans_table goes beyond PAGE_SIZE
Bugzilla: 1499411
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1499411

DESCRIPTION:
------------
On platforms with large number of Pstates, the transition table, which
is a NxN matrix, can overflow beyond the PAGE_SIZE boundary.

This can be seen on POWER9 which has 100+ Pstates.

As a result, each time the trans_table is read for any of the CPUs, we
will get the following error.

---------------------------------------------------
fill_read_buffer: show+0x0/0xa0 returned bad count
---------------------------------------------------

This patch ensures that in case of an overflow, we print a warning
once in the dmesg and return FILE TOO LARGE error for this and all
subsequent accesses of trans_table.

 Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>
 Acked-by: Viresh Kumar <viresh.kumar@linaro.org>
 Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>

BREW:
-----
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=14578408

TESTING:
--------
I have replicated the sosreport tests in a Power 9 system. After using a
kernel with this patch not only the bad count errors disappeared, but the
proper table exceeds message was shown.

UPSTREAM STATUS:
----------------
commit f7bc9b209e27c0b617378400136cc663a6314d0c

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=f7bc9b209e27c0b617378400136cc663a6314d0c

---------------
Desnes Augusto Nunes do Rosario

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>