[redhat] stop leaving empty temp files
Message-id: <20171123151930.23425-2-drjones@redhat.com>
Patchwork-id: 199914
O-Subject: [RHEL-ALT-7.5 PATCH 1/3] redhat: stop leaving empty temp files
Bugzilla:
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

I noticed several empty /tmp/tmp.* files after I had been running
'make rh-configs' several times. The problem is that the
rh-brew-test target in redhat/Makefile.test unconditionally
executes the RHS of the OUTFILE variable assignment, which
creates temp files, but then only removes them on the condition
that the target is made. This patch changes the creation of the
temp file to just the creation of the name. As that's racy, the
template is changed to further reduce potential for collisions.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>