[virt] KVM: arm/arm64: GICv4: Use the doorbell interrupt as an unblocking source
Message-id: <1512419424-15564-37-git-send-email-eric.auger@redhat.com>
Patchwork-id: 201096
O-Subject: [RHEL-ALT-7.5 BZ1386368 PATCH v2 36/44] KVM: arm/arm64: GICv4: Use the doorbell interrupt as an unblocking source
Bugzilla: 1386368
RH-Acked-by: Jeremy Linton <jlinton@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>

From: Marc Zyngier <marc.zyngier@arm.com>

The doorbell interrupt is only useful if the vcpu is blocked on WFI.
In all other cases, recieving a doorbell interrupt is just a waste
of cycles.

So let's only enable the doorbell if a vcpu is getting blocked,
and disable it when it is unblocked. This is very similar to
what we're doing for the background timer.

Reviewed-by: Christoffer Dall <cdall@linaro.org>
Reviewed-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>
(cherry picked from commit df9ba95993b9d59e077aa4613f7a4edd20a4064c)
Signed-off-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>