[cpufreq] powernv-cpufreq: Fix pstate_to_idx() to handle non-continguous pstates
Message-id: <20180117120408.10703.74906.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 203948
O-Subject: [PATCH RHEL-ALT-7.5 BZ1534472 2/3 v2] powernv-cpufreq: Fix pstate_to_idx() to handle non-continguous pstates
Bugzilla: 1534472
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1534472

Description:
------------
The code in powernv-cpufreq, makes the following two assumptions which
are not guaranteed by the device-tree bindings:

    1) Pstate ids are continguous: This is used in pstate_to_idx() to
       obtain the reverse map from a pstate to it's corresponding
       entry into the cpufreq frequency table.

    2) Every Pstate should always lie between the max and the min
       pstates that are explicitly reported in the device tree: This
       is used to determine whether a pstate reported by the PMSR is
       out of bounds.

Both these assumptions are unwarranted and can change on future
platforms.

In this patch, we maintain the reverse map from a pstate to it's index
in the cpufreq frequency table and use this in pstate_to_idx(). This
does away with the assumptions (1) mentioned above, and will work with
non continguous pstate ids. If no entry exists for a particular
pstate, then such a pstate is treated as being out of bounds. This
gets rid of assumption (2).

On all the existing platforms, where the pstates are 8-bit long
values, the new implementation of pstate_to_idx() takes constant time.

Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>

v2 - cleanup patch so it applies cleanly

Brew:
-----
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=15027225

Upstream Status:
----------------
commit 332f0a01f0dd669dcd208e4e9666d80dffd62e7b

https://git.kernel.org/pub/scm/linux/kernel/git/rafael/linux-pm.git/commit/?h=linux-next&id=332f0a01f0dd669dcd208e4e9666d80dffd62e7b

----------------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>