[netdrv] ibmvnic: Fix pending MAC address changes
Message-id: <20180117154234.12613.42494.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 203952
O-Subject: [PATCH RHEL-ALT-7.5 BZ1535360] ibmvnic: Fix pending MAC address changes
Bugzilla: 1535360
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1535360

Description:
------------
Due to architecture limitations, the IBM VNIC client driver is unable
to perform MAC address changes unless the device has "logged in" to
its backing device. Currently, pending MAC changes are handled before
login, resulting in an error and failure to change the MAC address.
Moving that chunk to the end of the ibmvnic_login function, when we are
sure that it was successful, fixes that.

The MAC address can be changed when the device is up or down, so
only check if the device is in a "PROBED" state before setting the
MAC address.

Fixes: c26eba03e407 ("ibmvnic: Update reset infrastructure to support tunable parameters")
Signed-off-by: Thomas Falcon <tlfalcon@linux.vnet.ibm.com>
Reviewed-by: John Allen <jallen@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Brew:
-----
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=15028526

Upstream:
---------
commit 3d1661304f0b2b51a8a43785b764822611dbdd53

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=3d1661304f0b2b51a8a43785b764822611dbdd53

Test Status:
------------
Successfully tested by IBM on power 9 system.

Red Hat doesn't have this hardware, so for a sanity test I made sure the
driver was able to load successfully.

[root@ibm-p9b-05 ibm]# insmod ./ibmvnic.ko

[ 1763.200493] ibmvnic: IBM System i/p Virtual NIC Driver 1.0.1

[root@ibm-p9b-05 ibm]# lsmod | more
Module                  Size  Used by
ibmvnic                99171  0

----------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>