[cpufreq] powernv-cpufreq: Treat pstates as opaque 8-bit values
Message-id: <20180119165033.39112.6901.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 204158
O-Subject: [PATCH RHEL-ALT-7.5 BZ1534472 3/3 v2] powernv-cpufreq: Treat pstates as opaque 8-bit values
Bugzilla: 1534472
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1534472

Description:
------------
On POWER8 and POWER9, the PMSR and the PMCR registers define pstates
to be 8-bit wide values. The device-tree exports pstates as 32-bit
wide values of which the lower byte is the actual pstate.

The current implementation in the kernel treats pstates as integer
type, since it used to use the sign of the pstate for performing some
boundary-checks. This is no longer required after the patch
"powernv-cpufreq: Fix pstate_to_idx() to handle non-continguous
pstates".

So, in this patch, we modify the powernv-cpufreq driver to uniformly
treat pstates as opaque 8-bit values obtained from the device-tree or
the PMCR. This simplifies the extract_pstate() helper function since
we no longer no longer require to worry about the sign-extentions.

Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>

v2 - cleanup so it now applies to current tree (kernel-alt-4.14.0-27.el7a)

Brew:
-----
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=15054015

Upstream Status:
----------------
commit 967b87fd81d513de7aac247320f02b1645f9ca64

https://git.kernel.org/pub/scm/linux/kernel/git/rafael/linux-pm.git/commit/?h=linux-next&id=967b87fd81d513de7aac247320f02b1645f9ca64

----------------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>