[netdrv] ibmvnic: Allocate and request vpd in init_resources
Message-id: <20180126181605.28794.64618.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 204422
O-Subject: [PATCH RHEL-ALT-7.5 BZ1537431 3/3] ibmvnic: Allocate and request vpd in init_resources
Bugzilla: 1537431
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1537431

Description:
------------
In reset events in which our memory allocations need to be reallocated,
VPD data is being freed, but never reallocated. This can cause issues if
we later attempt to access that memory or reset and attempt to free the
memory. This patch moves the allocation of the VPD data to init_resources
so that it will be symmetrically freed during release resources.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>
Reviewed-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Upstream Status:
----------------
commit 69d08dcbbe34347cbc044629cf6f25d062593dfe

https://git.kernel.org/pub/scm/linux/kernel/git/davem/net.git/commit/?id=69d08dcbbe34347cbc044629cf6f25d062593dfe

----------------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>