[powerpc] KVM: PPC: Book3S HV: Do SLB load/unload with guest LPCR value loaded
Message-id: <1517458786-9012-3-git-send-email-sbobroff@redhat.com>
Patchwork-id: 204663
O-Subject: [RHEL-ALT-7.5 KVM BZ1535753 PATCH 2/3] KVM: PPC: Book3S HV: Do SLB load/unload with guest LPCR value loaded
Bugzilla: 1535753
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>

From: Paul Mackerras <paulus@ozlabs.org>

This moves the code that loads and unloads the guest SLB values so that
it is done while the guest LPCR value is loaded in the LPCR register.
The reason for doing this is that on POWER9, the behaviour of the
slbmte instruction depends on the LPCR[UPRT] bit.  If UPRT is 1, as
it is for a radix host (or guest), the SLB index is truncated to
2 bits.  This means that for a HPT guest on a radix host, the SLB
was not being loaded correctly, causing the guest to crash.

The SLB is now loaded much later in the guest entry path, after the
LPCR is loaded, which for a secondary thread is after it sees that
the primary thread has switched the MMU to the guest.  The loop that
waits for the primary thread has a branch out to the exit code that
is taken if it sees that other threads have commenced exiting the
guest.  Since we have now not loaded the SLB at this point, we make
this path branch to a new label 'guest_bypass' and we move the SLB
unload code to before this label.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>

Signed-off-by: Sam Bobroff <sbobroff@redhat.com>
(cherry picked from commit 6964e6a4e4894c707e42d51d9d30683c57f43201)
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>