[netdrv] ibmvnic: Reset long term map ID counter
Message-id: <20180211161155.29462.12011.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 205606
O-Subject: [PATCH RHEL-ALT-7.5 BZ1544017 1/2] ibmvnic: Reset long term map ID counter
Bugzilla: 1544017
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1544017

Description:
------------
When allocating RX or TX buffer pools, the driver needs to provide a
unique mapping ID to firmware for each pool. This value is assigned
using a counter which is incremented after a new pool is created. The
ID can be an integer ranging from 1-255. When migrating to a device
that requests a different number of queues, this value was not being
reset properly. As a result, after enough migrations, the counter
exceeded the upper bound and pool creation failed. This is fixed by
resetting the counter to one in this case.

Signed-off-by: Thomas Falcon <tlfalcon@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Upstream Status:
----------------
commit faefaa97215a0c05105d7ae180fe1a3b5979ad1f

https://git.kernel.org/pub/scm/linux/kernel/git/davem/net.git/commit/?id=faefaa97215a0c05105d7ae180fe1a3b5979ad1f

----------------
Steve Best

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>