[s390] kernel: fix rwlock implementation
Message-id: <1519224313-10093-1-git-send-email-brueckner@redhat.com>
Patchwork-id: 205962
O-Subject: [RHEL ALT-7.5 PATCH] [s390] kernel: fix rwlock implementation
Bugzilla: 1547423
RH-Acked-by: Thomas Huth <thuth@redhat.com>
RH-Acked-by: Cornelia Huck <cohuck@redhat.com>

Description
-----------
A deadlock on an rwlock causes kernel hangs.  With upstream commit
94232a4332de ("s390/rwlock: improve writer fairness") rwlock writer
fairness was supposed to be implemented. If a writer tries to take
an rwlock it sets unconditionally the writer bit within the lock word
and waits until all readers have released the lock. This however can
lead to a deadlock since rwlocks can be taken recursively by readers.

If e.g. CPU 0 holds the lock as a reader, and CPU 1 wants to write-lock
the lock, then CPU 1 sets the writer bit and afterwards busy waits for
CPU 0 to release the lock. If now CPU 0 tries to read-lock the lock
again (recursively) it will also busy wait until CPU 1 removes the
writer bit, which will never happen, since it waits for the first
reader on CPU 0 to release the lock.

Revert the rwlock writer fairness semantics again.

Bugzilla
--------
BZ 1547423
https://bugzilla.redhat.com/show_bug.cgi?id=1547423

Upstream status of the patch
----------------------------
This patch is specific for kernel v4.14.  With v4.15, the problem is
indirectly corrected with the introduction of queued spinlocks.

Brew
----
https://brewweb.devel.redhat.com/taskinfo?taskID=15360893

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>