proc/kcore: don't grab lock for kclist_add()
kclist_add() is only called at init time, so there's no point in
grabbing any locks. We're also going to replace the rwlock with a rwsem,
which we don't want to try grabbing during early boot.

While we're here, mark kclist_add() with __init so that we'll get a
warning if it's called from non-init code.

Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Omar Sandoval <osandov@fb.com>